Changing routes to fit excel.
