+++ src/api/routes/index.js
@@ -2,10 +2,9 @@ import { Router } from 'express';
 import swaggerJsdoc from 'swagger-jsdoc';
 import { serve, setup } from 'swagger-ui-express';
 import { specs, swaggerConfig } from '../../config/index.js';
+import auth from './auth.js';
 import user from './user.js';
-import category from './category.js';
-import recipe from './recipe.js';
-import ingredient from './ingredient.js';
+import recipes from './recipes.js';
 const router = Router();
 
 const specDoc = swaggerJsdoc(swaggerConfig);
@@ -13,9 +12,8 @@ const specDoc = swaggerJsdoc(swaggerConfig);
 router.use(specs, serve);
 router.get(specs, setup(specDoc, { explorer: true }));
 
+router.use('/auth', auth);
 router.use('/user', user);
-router.use('/category', category);
-router.use('/recipe', recipe);
-router.use('/ingredient', ingredient);
+router.use('/recipes', recipes);
 
 export default router;
